requirements.txt
sentence-transformers
faiss-cpu
transformers
pypdf
gradio
scikit-learn
numpy
pandas